pyproject.toml
[tool.poetry]
name = "contact-book-api"
version = "0.1.0"
description = ""
authors = ["Denis Volokh <denis.volokh@gmail.com>"]
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.10"
mkdocs = "^1.5.2"
fastapi = "^0.101.0"
pydantic = "^2.1.1"
databases = "^0.7.0"
asyncpg = "^0.28.0"
uvicorn = "^0.23.2"
psycopg2-binary = "^2.9.7"
huey = "^2.4.5"
redis = "^4.6.0"
celery = "^5.3.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
black = "^23.7.0"
mypy = "^1.5.1"
pyproject-flake8 = "^6.0.0.post1"
types-requests = "^2.31.0.2"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output=3
include_trailing_comma = true
force_grid_wrap=0
use_parentheses = true
line_length=88

[tool.flake8]
ignore = "E203, E266, E501, W503"
max-line-length = 88
max-complexity = 18
select = "B,C,E,F,W,T4"

[tool.mypy]
files="api/*.py"
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true